feat(src): add branch protections and codeowners
Expanded the settings configured for `./.github/settings.yml` by adding
in branch protections, vulnerability scanning, and automatic security
fixes.

Also adds a `CODEOWNERS` file which sets myself as the global codeowner,
this is merely a precaution.

fix(src): admins bypass verifications

Allows users with the bypass branch protections permission to bypass the
main branch protections.

This is due to the fact that I am the sole contributor for most of my
repositories and will need to merge PRs without review from another
party.